requirements.txt
streamlit
numpy
pandas
opencv-python
keras
tensorflow
matplotlib
streamlit-drawable-canvas
gdown
Pillow
